deepfake-audio-detector--main/requirements.txt
streamlit
librosa
numpy
matplotlib
tensorflow
scipy
praat-parselmouth
soundfile
scikit-learn
